+++ App.tsx
@@ -1,6 +1,9 @@
-import React, { useState, useEffect } from 'react';
-import { StyleSheet, Text, View, Button } from 'react-native';
+import React, { useState } from 'react';
+import { Text, View, Button } from 'react-native';
 import MatterhornApiClient, { CounterResponse } from './api-client/MatterhornApiClient';
+import PocButton from './poc/components/PocButton';
+import Spacer from './poc/components/Spacer';
+import styles from './poc/styles';
 
 const apiBaseUrl = 'http://matterhornapiservice-env.eba-qjezc5kq.us-east-1.elasticbeanstalk.com';
 const apiClient: MatterhornApiClient = new MatterhornApiClient(apiBaseUrl);
@@ -15,19 +18,12 @@ export default function App() {
     <View style={styles.container}>
       <Text>{message}</Text>
       <Text>Counter: {count}</Text>
-      <Button onPress={() => retrieveCount(apiClient, setCount)} title="Click me"/>
+      <Spacer width={50} height={50} />
+      <PocButton title='Click Me' onPress={() => retrieveCount(apiClient, setCount)} />
     </View>
   );
 }
 
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#fff',
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-});
 
 function retrieveCount(api: MatterhornApiClient, setCount: React.Dispatch<React.SetStateAction<number>>) {
   api.Counter().then((value: CounterResponse) => {
