+++ poc/components/Spacer.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+import { View } from 'react-native';
+
+
+export default function Spacer(props) {
+    return (
+        <View style={{width: props.width ?? 0, height: props.height ?? 0}}/>
+    );
+}
