+++ poc/styles.ts
@@ -1,0 +1,26 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+    container: {
+        flex: 4,
+        backgroundColor: '#fff',
+        alignItems: 'center',
+        justifyContent: 'center',
+    },
+    button: {
+        width: 150,
+        height: 50,
+        backgroundColor: 'skyblue'
+    },
+    buttonText: {
+        textAlign: 'center',
+        color: 'white',
+        padding: 15
+    },
+    spacer: {
+        width: 50,
+        height: 50
+    }
+});
+
+export default styles;
